+++ testcases/kernel/syscalls/msync/msync02.c
@@ -140,7 +140,7 @@ void setup(void)
 		tst_brkm(TBROK | TERRNO, cleanup, "lseek failed");
 
 	/* Write the string in write_buf at the 100 byte offset */
-	if (write(fildes, write_buf, strlen(write_buf)) != strlen(write_buf))
+	if (write(fildes, write_buf, strlen(write_buf)) != (long)strlen(write_buf))
 		tst_brkm(TBROK | TERRNO, cleanup, "write failed");
 }
