+++ testcases/kernel/syscalls/setrlimit/setrlimit01.c
@@ -151,8 +151,7 @@ static void test2(void)
 
 		bytes = write(fd, buf, 26);
 		if (bytes != 10) {
-			if (write(pipefd[1], &bytes, sizeof(bytes))
-			    < sizeof(bytes)) {
+			if (write(pipefd[1], &bytes, sizeof(bytes)) < (long)sizeof(bytes)) {
 				perror("child: write to pipe failed");
 			}
 			close(pipefd[1]);	/* EOF */
@@ -177,7 +176,7 @@ static void test2(void)
 		break;
 	case 3:
 		close(pipefd[1]);	/* close unused write end */
-		if (read(pipefd[0], &bytes, sizeof(bytes)) < sizeof(bytes))
+		if (read(pipefd[0], &bytes, sizeof(bytes)) < (long)sizeof(bytes))
 			tst_resm(TFAIL, "parent: reading pipe failed");
 
 		close(pipefd[0]);
